SYS-461 Refactor Money handling to use integer amounts, update schemas and tests accordingly

+++ src/types/DebtType.ts
@@ -4,7 +4,7 @@ export const BalanceSchema = z.object({
     playerId: z.number().int().min(1).nullable(),
     maxGameDayId: z.number().int().min(1).optional(),
     playerName: z.string().min(1),
-    amount: z.number(),
+    amount: z.number().int(),
 });
 
 export type BalanceType = z.infer<typeof BalanceSchema>;
@@ -24,9 +24,9 @@ export type ClubBalanceType = z.infer<typeof ClubBalanceSchema>;
 export const BalanceSummarySchema = z.object({
     players: z.array(PlayerBalanceSchema),
     club: ClubBalanceSchema,
-    total: z.number(),
-    positiveTotal: z.number().nonnegative(),
-    negativeTotal: z.number().nonpositive(),
+    total: z.number().int(),
+    positiveTotal: z.number().int().nonnegative(),
+    negativeTotal: z.number().int().nonpositive(),
 });
 
 export type BalanceSummaryType = z.infer<typeof BalanceSummarySchema>;
